+++ .devcontainer/post-create.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+# Yarn repo causes GPG varification issues
+sudo rm -f /etc/apt/sources.list.d/yarn.list
+
 sudo apt update
 sudo apt install -y vim libudunits2-dev plantuml